- revised datatype for stored procedure - removed troubleshooting section for Dashboard because it takes too long to load

+++ BluePrints/EntityFramework/BluePrintsData/X_EARNED_QUERY.cs
@@ -27,7 +27,7 @@ namespace BluePrints.Data
         [StringLength(100)]
         public string VariationCode { get; set; }
 
-        public decimal? RATE { get; set; }
+        public double? RATE { get; set; }
 
         public Guid GUID { get; set; }
